requirements.txt
torch
torchvision
streamlit
numpy
Pillow
torchsummary
matplotlib